+++ skills/chatgpt-app-builder/references/ecommerce.md
@@ -2,7 +2,7 @@
 
 The `ecom` template is a skeleton: the wiring is in place, the data is not. One tool, `search-products`, takes a keyword plus filters, runs a `search()` stub, and narrates the results for the model. This reference connects it to a real catalog.
 
-Not in a scaffolded `ecom` project (no `search-products` tool in `src/server.ts`)? Scaffold it first with the `--ecom` flag: [copy-template.md](copy-template.md). Then return here.
+Not in a scaffolded `ecom` project (no `src/tools/search-products.ts`)? Scaffold it first with the `--ecom` flag: [copy-template.md](copy-template.md). Then return here.
 
 Work in the order below. Do not skip ahead: each step depends on the previous ones. Never invent a schema, endpoint, or credential.
 
@@ -21,7 +21,9 @@ Using the docs and credentials, run real queries. Confirm the exact response sha
 
 ## 4. Fill the @todos and verify
 
-`grep -rn "@todo" src`, then resolve every marker. Fill the server-wide items, then each tool: complete its `@todo`s and verify it before moving to the next.
+Layout: `src/config.ts` (shared tuning), `src/server.ts` (name/version, server prompt, tool registration), and one file per tool under `src/tools/`.
+
+`grep -rn "@todo" src`, then resolve every marker. Fill the shared items, then each tool: complete its `@todo`s and verify it before moving to the next.
 
 To verify: `pnpm build` (typecheck), then start the dev server and keep it running:
 
@@ -31,15 +33,15 @@ To verify: `pnpm build` (typecheck), then start the dev server and keep it runni
 
 Each tool's verify curl is in its section below.
 
-### Server-wide (`src/server.ts`, plus `.env.template`)
+### Shared
 
 - [ ] `.env.template`: regenerate from the user's `.env` (same keys, values blanked) so the committed placeholder matches the credentials the integration needs.
-- [ ] `name` / `version`.
-- [ ] server `instructions`: adapt the server-wide prompt to the catalog.
-- [ ] `CAROUSEL_MAX_SIZE`: number of products to curate toward.
-- [ ] `MIN_SEARCH_ITERATIONS`: minimum searches before rendering.
+- [ ] `src/server.ts` — `name` / `version`.
+- [ ] `src/server.ts` — server `instructions`: adapt the server-wide prompt to the catalog.
+- [ ] `src/config.ts` — `CAROUSEL_MAX_SIZE`: number of products to curate toward.
+- [ ] `src/config.ts` — `MIN_SEARCH_ITERATIONS`: minimum searches before rendering.
 
-### `search-products`
+### `search-products` (`src/tools/search-products.ts`)
 
 - [ ] `description`: describe the catalog, its categories, and the search/curate loop.
 - [ ] `_meta`: the invoking and invoked status messages.
@@ -47,6 +49,7 @@ Each tool's verify curl is in its section below.
 - [ ] `inputSchema.sort`: set the real sort options, or remove.
 - [ ] `inputSchema` filters: replace `priceRange` with one optional param per real facet.
 - [ ] `search()`: query the data source with `keyword`, `sort`, and filters; map each hit to `{ id, properties: [{ name, value: string[] }] }[]`; set `pages` and `totalHits`.
+- [ ] `narrate()` NEXT STEP: adapt the post-search guidance to your flow.
 
 Verify: curl the stateless `/mcp` endpoint (`Accept` must include `text/event-stream` or the SDK rejects the request). Set `arguments` to `keyword` plus any `sort`/filters you implemented:
